+++ src/test/java/stroom/dashboard/expression/v1/TestExpressionParser.java
@@ -67,6 +67,7 @@ public class TestExpressionParser {
         test("1=0");
         test("decode('fred', 'fr.+', 'freda', 'freddy')");
         test("extractHostFromUri('http://www.example.com:1234/this/is/a/path')");
+        test("link('title', 'http://www.somehost.com', 'somepath', 'target')");
     }
 
     private void test(final String expression) throws ParseException {
@@ -508,6 +509,16 @@ public class TestExpressionParser {
 
         final Val out = gen.eval();
         Assert.assertEquals("hello", out.toString());
+    }
+
+    @Test
+    public void testLink() throws ParseException {
+        final Generator gen = createGenerator("link('Title', 'http://www.somehost.com', '/somepath')");
+
+        gen.set(getVal("this"));
+
+        final Val out = gen.eval();
+        Assert.assertEquals("[Title](http://www.somehost.com/somepath){BROWSER_TAB}", out.toString());
     }
 
     @Test